requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
Faker==25.9.2
pytest==8.2.2
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
requests==2.32.3
selenium==4.21.0


